enhanced_group_memory/requirements.txt
streamlit==1.29.0
textblob==0.17.1
plotly==5.18.0
pandas>=2.1.0
numpy>=1.24.0
sentence-transformers==2.2.2
scikit-learn>=1.3.0
huggingface-hub==0.19.4
python-dotenv==1.0.0
